move TestRunnerConfig inner class out of InstrumentationTestsTask and rename it to TestRunnersConfig it is big enough to stand on its own
