+++ src/main/groovy/com/jvoegele/gradle/tasks/android/InstrumentationTestsTask.groovy
@@ -4,79 +4,8 @@ import com.jvoegele.gradle.tasks.android.exceptions.InstrumentationTestsFailedEx
 import com.jvoegele.gradle.tasks.android.exceptions.AdbErrorException;
 
 class InstrumentationTestsTask extends AdbExec {
-
-  /**
-   * Configures the test runners used to run the instrumentation tests.
-   *  
-   * @author Matthias Kaeppler
-   */
-  private class TestRunnerConfig {
-    
-    static final String PACKAGE    = "testpackage"
-    static final String ANNOTATION = "annotation"
-    static final String RUNNER     = "with"
-    static final String NAME       = "name"
-    static final String OPTIONS    = "options"
-    
-    def defaultConfig
-    def packageRunners    = [:]
-    def annotationRunners = [:]
-    
-    def run(args = [:]) {
-      createRunConfig(args)
-    }
-    
-    def createRunConfig(args = [:]) {
-      def testRunner  = args[(RUNNER)] ?: getProject().convention.plugins.android.testRunner
-      def packageName = args[(PACKAGE)]
-      def annotation  = args[(ANNOTATION)]
-      def name        = args[(NAME)] ?: "instrumentation-tests-$numRunners"
-      def options     = args[(OPTIONS)] ?: []
-
-      if (options instanceof String || options instanceof GString) {
-        options = [options]
-      }
-      
-      // always wait for tests to finish
-      options << "-w"
-      // enable support for Zutubi's JUnit report test runner
-      options << "-e reportFilePath ${name}.xml"
-           
-      if (packageName) {
-        packageRunners[expandPackageName(packageName)] = buildRunner(testRunner, name, options)
-      } else if (annotation) {
-        annotationRunners[annotation] = buildRunner(testRunner, name, options)
-      } else {
-        defaultConfig = buildRunner(testRunner, name, options)
-      }
-    }
-    
-    boolean performDefaultRun() {
-      packageRunners.isEmpty() && annotationRunners.isEmpty()
-    }
-    
-    int getNumRunners() {
-      packageRunners.size() + annotationRunners.size()
-    }
-    
-    def buildRunner(def testRunner, def name, def options) {
-      def wrapper = new Expando()
-      wrapper.runner = testRunner
-      wrapper.name = name
-      wrapper.options = options
-      wrapper
-    }
-    
-    String expandPackageName(String packageName) {
-      if (!packageName.startsWith(getTestPackage())) {
-        return "${getTestPackage()}.${packageName}"
-      }
-      return packageName;
-    }
-  }
-  
   def testPackage
-  def testRunnerConfig
+  def testRunnersConfig
   def testReportsSourcePath
   def testReportsTargetPath
   def defaultAdbArgs
@@ -87,7 +16,7 @@ class InstrumentationTestsTask extends AdbExec {
     this.testPackage = ant['manifest.package']
     this.testedPackage = ant['tested.manifest.package']
     this.defaultAdbArgs = super.getArgs()
-    this.testRunnerConfig = new TestRunnerConfig()
+    this.testRunnersConfig = new TestRunnersConfig(project, testPackage)
     
     if (testedPackage) { // this is only set for instrumentation projects
       this.testReportsSourcePath = "/data/data/$testedPackage/files"
@@ -116,29 +45,29 @@ class InstrumentationTestsTask extends AdbExec {
    * @return
    */
   def runners(Closure config) {
-    config.delegate = testRunnerConfig
+    config.delegate = testRunnersConfig
     config()
   }
   
   @Override
   def exec() {
 
-    if (testRunnerConfig.performDefaultRun()) {
-      def runConfig = testRunnerConfig.defaultConfig 
+    if (testRunnersConfig.performDefaultRun()) {
+      def runConfig = testRunnersConfig.defaultConfig
       if (!runConfig) {
         // this will happen if there was no runners block provided at all
-        runConfig = testRunnerConfig.createRunConfig()
+        runConfig = testRunnersConfig.createRunConfig()
       }
       performTestRun(runConfig)
     } else {
       // execute package specific runners
-      testRunnerConfig.packageRunners.each {
+      testRunnersConfig.packageRunners.each {
         def packageName = it.key
         def runConfig   = it.value
         performTestRun(runConfig, "-e package $packageName")
       }
       // execute annotation specific runners
-      testRunnerConfig.annotationRunners.each {
+      testRunnersConfig.annotationRunners.each {
         def annotation = it.key
         def runConfig = it.value
         performTestRun(runConfig, "-e annotation $annotation")
